Separated Json Object/Serialization Code into file
