+++ beacon_node/execution_layer/src/test_utils/handle_rpc.rs
@@ -1,5 +1,6 @@
 use super::Context;
 use crate::engine_api::{http::*, ExecutePayloadResponse, ExecutePayloadResponseStatus};
+use crate::json_structures::*;
 use serde::de::DeserializeOwned;
 use serde_json::Value as JsonValue;
 use std::sync::Arc;
